devide app into a different components(header.vue, mobileSideBar.vue, overlay.vue,footer.vue)
