+++ src/components/footer.vue
@@ -1,0 +1,12 @@
+<script setup>
+</script>
+
+<template>
+	<footer id="footer">
+		<div class="w-full z-20 ">
+			<div class="">
+				
+			</div>
+		</div>
+	</footer>
+</template>
