+++ src/views/ToolsView.vue
@@ -1,6 +1,23 @@
 <script setup>
+import Header from '../components/header.vue';
+import Footer from '../components/footer.vue';
+import MobileSideBar from '../components/mobileSideBar.vue';
+import Overlay from '../components/overlay.vue';
 </script>
 
 <template>
-	Hi there
+	<!-- Header -->
+	<Header />
+
+	<!-- Mobile SideBar -->
+	<MobileSideBar />		
+
+	<!-- Overlay -->
+	<Overlay />
+	<main>
+		Hey
+	</main>
+
+	<!-- Footer -->
+	<Footer />
 </template>
